+++ src/cobalt/renderer/glimp_shaders/glsl/fragment_textured_vbo_uyvy_1plane.glsl
@@ -30,5 +30,6 @@ void main() {
   }
 
   vec4 untransformed_color = vec4(y_value, uv_value.r, uv_value.g, 1.0);
-  gl_FragColor = untransformed_color * to_rgb_color_matrix;
+  vec4 color = untransformed_color * to_rgb_color_matrix;
+  gl_FragColor = color;
 }