+++ src/cobalt/renderer/glimp_shaders/glsl/fragment_textured_vbo_yuv_3plane.glsl
@@ -11,5 +11,6 @@ void main() {
     texture2D(texture_y, v_tex_coord_y).a,
     texture2D(texture_u, v_tex_coord_u).a,
     texture2D(texture_v, v_tex_coord_v).a, 1.0);
-  gl_FragColor = untransformed_color * to_rgb_color_matrix;
+  vec4 color = untransformed_color * to_rgb_color_matrix;
+  gl_FragColor = color;
 }