+++ src/cobalt/renderer/rasterizer/egl/textured_mesh_renderer.cc
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include "base/strings/stringprintf.h"
+#include "cobalt/extension/graphics.h"
 #include "cobalt/math/size.h"
 #include "cobalt/renderer/backend/egl/utils.h"
 #include "cobalt/renderer/egl_and_gles.h"
@@ -346,12 +347,53 @@ uint32 TexturedMeshRenderer::CreateFragmentShader(
   if (color_matrix) {
     blit_fragment_shader_source +=
         ");"
-        "  gl_FragColor = untransformed_color * to_rgb_color_matrix;"
-        "}";
+        "  vec4 color = untransformed_color * to_rgb_color_matrix;";
   } else {
     blit_fragment_shader_source +=
         ");"
-        "  gl_FragColor = untransformed_color;"
+        "  vec4 color = untransformed_color;";
+  }
+#if SB_API_VERSION >= 11
+  const CobaltExtensionGraphicsApi* graphics_extension =
+      static_cast<const CobaltExtensionGraphicsApi*>(
+          SbSystemGetExtension(kCobaltExtensionGraphicsName));
+  CobaltExtensionGraphicsMapToMeshColorAdjustment color_adjustment;
+  if (graphics_extension != nullptr &&
+      strcmp(graphics_extension->name, kCobaltExtensionGraphicsName) == 0 &&
+      graphics_extension->version >= 5 &&
+      graphics_extension->GetMapToMeshColorAdjustments(&color_adjustment)) {
+    // Setup vectors for the color adjustments. Ensure they use dot for decimal
+    // point regardless of locale.
+    std::stringstream buffer;
+    buffer.imbue(std::locale::classic());
+    buffer << "  vec4 scale0 = vec4(" << color_adjustment.rgba0_scale[0] << ","
+           << color_adjustment.rgba0_scale[1] << ","
+           << color_adjustment.rgba0_scale[2] << ","
+           << color_adjustment.rgba0_scale[3] << ");";
+    buffer << "  vec4 scale1 = vec4(" << color_adjustment.rgba1_scale[0] << ","
+           << color_adjustment.rgba1_scale[1] << ","
+           << color_adjustment.rgba1_scale[2] << ","
+           << color_adjustment.rgba1_scale[3] << ");";
+    buffer << "  vec4 scale2 = vec4(" << color_adjustment.rgba2_scale[0] << ","
+           << color_adjustment.rgba2_scale[1] << ","
+           << color_adjustment.rgba2_scale[2] << ","
+           << color_adjustment.rgba2_scale[3] << ");";
+    buffer << "  vec4 scale3 = vec4(" << color_adjustment.rgba3_scale[0] << ","
+           << color_adjustment.rgba3_scale[1] << ","
+           << color_adjustment.rgba3_scale[2] << ","
+           << color_adjustment.rgba3_scale[3] << ");";
+    blit_fragment_shader_source +=
+        "  vec4 color2 = color * color;"
+        "  vec4 color3 = color2 * color;" +
+        buffer.str() +
+        "  color = scale0 + scale1*color + scale2*color2 + scale3*color3;"
+        "  gl_FragColor = clamp(color, vec4(0.0), vec4(1.0));"
+        "}";
+  } else
+#endif  // SB_API_VERSION >= 11
+  {
+    blit_fragment_shader_source +=
+        "  gl_FragColor = color;"
         "}";
   }